refatoracao nos servicos de tatuadores

+++ src/main/java/br/com/legacyink/domain/service/EstudioService.java
@@ -1,10 +1,10 @@
 package br.com.legacyink.domain.service;
 
+import br.com.legacyink.domain.exception.EntidadeEmUsoException;
 import br.com.legacyink.domain.exception.EstudioNaoEncontradoException;
 import br.com.legacyink.domain.model.Estudio;
 import br.com.legacyink.domain.repository.EstudioRepository;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import javax.persistence.Transient;
@@ -13,12 +13,16 @@ import java.util.List;
 @Service
 public class EstudioService {
 
+    private final EstudioRepository estudioRepository;
+
     @Autowired
-    private EstudioRepository estudioRepository;
+    public EstudioService(EstudioRepository estudioRepository) {
+        this.estudioRepository = estudioRepository;
+    }
 
     public Estudio buscaEstudioOuErro(Long id) {
         return estudioRepository.findById(id).orElseThrow(
-                () -> new RuntimeException("Estudio não encontrado"));
+                () -> new EstudioNaoEncontradoException(String.format("Estudio não encontrado: %d", id)));
     }
 
     public List<Estudio> listarEstudios() {
@@ -32,10 +36,20 @@ public class EstudioService {
 
     @Transient
     public void remover(Long id) {
-        try {
-            estudioRepository.deleteById(id);
-        } catch (EmptyResultDataAccessException e) {
-            throw new EstudioNaoEncontradoException(id);
+        Estudio estudio = estudioRepository.findById(id)
+                .orElseThrow(() -> new EstudioNaoEncontradoException(id));
+
+        if (!estudio.getEstoque().isEmpty()) {
+            throw new EntidadeEmUsoException("O Estudio possui itens no estoque");
         }
+
+        if (!estudio.getTatuadores().isEmpty()) {
+            throw new EntidadeEmUsoException("O Estudio possui tatuadores associados");
+        }
+
+        if (!estudio.getClientes().isEmpty()) {
+            throw new EntidadeEmUsoException("O Estudio possui clientes associados");
+        }
+        estudioRepository.deleteById(id);
     }
 }
